+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Next version
-
+- Add download of all invoices in pdf
 - Rename `render_dorsale_index` to `render_dorsale_page`
 - Replace `render_contextual` by `render_dorsale_page`
 - Change some view/partial names